docker addition
+add docker-compose.yml
+add nginx (test version size_fix_nginx.conf)

***
size_fix_nginx.conf - is a finalized version of the nginx.conf file. But for some reasons it was not put into production because it was not needed, but left in case of quick migration of configuration in case of need.
***
The entire commit is only designed to launch port 80. For the 443 port there will be further modification of the configuration at the final stage of implementation.